requirements.txt
web3==6.15.1
mnemonic==0.20
fpdf2==2.7.6
segno==1.5.2 